Switch list.notes from being an array to a function
This is required because if list.notes is an array for the special lists "all" and "fav", they are evaluated only at application start, and not at every call. Thus, if user favorites or unfavorites different notes, or adds new notes, they do not show up since the notes array is not reevaluated. But with a function call, it will be evaluated every time it is called.

Since regular lists will not use any evaluation dependent logic, they don't really need a function, but in order to have a consistent interface for lists, we switched them over to being functions as well, which trivially return an underlying array.